+++ src/Field/Field.jsx
@@ -82,9 +82,12 @@ class Field extends React.Component {
 
   renderLabel(options = {}) {
     const t = this;
+    const { errMsg, showErrInTip } = this.props
+    const hasError = errMsg && showErrInTip && t.props.layout === 'v';
     if (t.props.label || t.props.labelRight) {
       return (
         <Label
+          className={hasError ? prefixClass('field-label-error') : ''}
           label={t.props.label}
           required={t.props.required}
           layout={t.props.layout}
@@ -141,6 +144,7 @@ class Field extends React.Component {
           [t.props.className]: !!t.props.className,
           [prefixClass('FBH FBAC')]: middleIcon
         })}
+        onClick={(e) => {this.props.onClick(e)}}
       >
         <div className={classnames({
           [prefixClass('field-pos-box')]: middleIcon,
@@ -172,6 +176,7 @@ Field.defaultProps = {
   toastComponent: undefined,
   errMsg: undefined,
   showErrInTip: true,
+  onClick: () => {}
 };
 
 // http://facebook.github.io/react/docs/reusable-components.html
@@ -197,6 +202,7 @@ Field.propTypes = {
   toastComponent: PropTypes.func,
   errMsg: PropTypes.string,
   showErrInTip: PropTypes.bool,
+  onClick: PropTypes.func
 };
 
 Field.getFieldProps = (props = {}) => {