+++ src/TextField/TextField.jsx
@@ -52,7 +52,7 @@ class TextField extends React.Component {
 
   renderInput() {
     const t = this;
-    const { readOnly } = t.props;
+    const { readOnly, renderView } = t.props;
     return (
       <div className={prefixClass('text-field-content-main')}>
         <div
@@ -73,7 +73,7 @@ class TextField extends React.Component {
             onFocus={(e) => { t.handleFocus(e); }}
             onBlur={(e) => { t.handleBlur(e); }}
           />
-        ) : <span>{t.props.value || '　'}</span>}
+        ) : <span>{renderView ? renderView(t.props.value) : (t.props.value || '　')}</span>}
       </div>
     );
   }
@@ -133,6 +133,7 @@ TextField.defaultProps = {
   value: '',
   allowClear: true,
   children: undefined,
+  renderView: undefined
 };
 
 TextField.propTypes = {
@@ -149,6 +150,7 @@ TextField.propTypes = {
   children: PropTypes.any,
   allowClear: PropTypes.bool,
   disabled: PropTypes.bool,
+  renderView: PropTypes.func
 };
 
 TextField.displayName = 'TextField';